New gen-as-const-headers infinity-nptl{,_db}-constants.h
This commit adds two new generated assembler constants headers to
export constants used while building libpthread.so and
libthread_db.so.  It also adds a rule to build them to nptl/Makefile.

I would prefer to have infinity-nptl_db-constants.sym in nptl_db
and have it built by nptl_db/Makefile but the ordering of the build
seems to make this impossible: nptl_db needs to be built after nptl,
but infinity-nptl_db-constants.h is required to build the infinity
notes for libpthread.so.  If it's possible to change this then I
will.  In the meantime I had to add "breadcrumb" header files for
nptl_db's proc_service.h and thread_db.h for use during the build.